# backend/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "BatchJobsResponse" : {
        "properties" : {
          "completed" : {
            "format" : "int32",
            "type" : "integer"
          },
          "failed" : {
            "format" : "int32",
            "type" : "integer"
          },
          "running" : {
            "format" : "int32",
            "type" : "integer"
          },
          "stopped" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "completed", "failed", "running", "stopped" ],
        "type" : "object"
      },
...
      "JobRequestResponse" : {
        "properties" : {
          "finishedAt" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "jobExecutionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "jobName" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "parameters" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "type" : "object"
          },
          "requestedAt" : {
            "type" : "string"
          },
          "requestedBy" : {
            "type" : "string"
          },
          "runId" : {
            "type" : "string"
          },
          "startedAt" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "targetRunId" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "jobName", "kind", "parameters", "requestedAt", "requestedBy", "status" ],
        "type" : "object"
      },
      "JobRunDetailResponse" : {
        "properties" : {
          "batchCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "batches" : {
            "items" : {
              "$ref" : "#/components/schemas/StreamBatchResponse"
            },
            "type" : "array"
          },
          "duplicateCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "durationMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "endedAt" : {
            "type" : "string"
          },
          "exitCode" : {
            "type" : "string"
          },
          "exitMessage" : {
            "type" : "string"
          },
          "jobExecutionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "type" : "string"
          },
          "links" : {
            "$ref" : "#/components/schemas/LinksResponse"
          },
          "name" : {
            "type" : "string"
          },
          "parameters" : {
            "items" : {
              "$ref" : "#/components/schemas/ParameterResponse"
            },
            "type" : "array"
          },
          "readCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "request" : {
            "$ref" : "#/components/schemas/RequestResponse"
          },
          "restartable" : {
            "type" : "boolean"
          },
          "runId" : {
            "type" : "string"
          },
          "skipCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "startedAt" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "steps" : {
            "items" : {
              "$ref" : "#/components/schemas/StepResponse"
            },
            "type" : "array"
          },
          "writeCount" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "kind", "name", "runId", "status" ],
        "type" : "object"
      },
      "JobRunResponse" : {
        "properties" : {
          "batchCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "batchIds" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "duplicateCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "durationMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "endedAt" : {
            "type" : "string"
          },
          "exitCode" : {
            "type" : "string"
          },
          "exitMessage" : {
            "type" : "string"
          },
          "jobExecutionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "readCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "request" : {
            "$ref" : "#/components/schemas/RequestResponse"
          },
          "restartable" : {
            "type" : "boolean"
          },
          "runId" : {
            "type" : "string"
          },
          "skipCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "startedAt" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "writeCount" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "kind", "name", "readCount", "runId", "skipCount", "status", "writeCount" ],
        "type" : "object"
      },
      "JobSummaryResponse" : {
        "properties" : {
          "batchJobs" : {
            "$ref" : "#/components/schemas/BatchJobsResponse"
          },
          "bucket" : {
            "type" : "string"
          },
          "from" : {
            "type" : "string"
          },
          "stream" : {
            "items" : {
              "$ref" : "#/components/schemas/SourceSeriesResponse"
            },
            "type" : "array"
          },
          "to" : {
            "type" : "string"
          }
        },
        "required" : [ "batchJobs", "bucket", "from", "stream", "to" ],
        "type" : "object"
      },
...
      "PagedResponseJobRunResponse" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/JobRunResponse"
            },
            "type" : "array"
          },
          "nextCursor" : {
            "type" : "string"
          }
        },
        "required" : [ "items" ],
        "type" : "object"
      },
...
      "Problem" : {
        "additionalProperties" : true,
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "errors" : {
            "items" : {
              "properties" : {
                "field" : {
                  "type" : "string"
                },
                "message" : {
                  "type" : "string"
                }
              },
              "type" : "object"
            },
            "type" : "array"
          },
          "instance" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          },
          "type" : {
            "enum" : [ "urn:pti:problem:validation-error", "urn:pti:problem:unauthorized", "urn:pti:problem:forbidden", "urn:pti:problem:not-found", "urn:pti:problem:method-not-allowed", "urn:pti:problem:not-acceptable", "urn:pti:problem:conflict", "urn:pti:problem:dlq-invalid-state", "urn:pti:problem:replay-already-running", "urn:pti:problem:job-not-restartable", "urn:pti:problem:job-not-running", "urn:pti:problem:payload-too-large", "urn:pti:problem:unsupported-media-type", "urn:pti:problem:invalid-payload", "urn:pti:problem:pii-not-allowed", "urn:pti:problem:business-key-changed", "urn:pti:problem:replay-window-invalid", "urn:pti:problem:unsupported-source", "urn:pti:problem:analytics-recompute-unavailable", "urn:pti:problem:idempotency-key-reused", "urn:pti:problem:job-not-allowed", "urn:pti:problem:invalid-flag-value", "urn:pti:problem:rate-limited", "urn:pti:problem:internal-error", "urn:pti:problem:simulator-unavailable", "urn:pti:problem:service-unavailable" ],
            "type" : "string"
          }
        },
        "required" : [ "status", "title", "traceId", "type" ],
        "type" : "object"
      },
...
      "RunJobBody" : {
        "properties" : {
          "jobName" : {
            "minLength" : 1,
            "type" : "string"
          },
          "parameters" : {
            "additionalProperties" : { },
            "type" : "object"
          }
        },
        "required" : [ "jobName" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "A Keycloak access token of realm pti.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Transit, insight, alert and ETL operations API of the PTI platform.",
    "title" : "Public Transport Intelligence API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/etl/jobs" : {
      "get" : {
        "operationId" : "listJobRuns",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "kind",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "name",
          "required" : false,
          "schema" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "runs" : {
                    "description" : "runs",
                    "value" : {
                      "items" : [ {
                        "batchIds" : [ "0192f5a1-7c1e-7d3a-9b2c-4e5f6a7b8c9d" ],
                        "durationMs" : 729000,
                        "endedAt" : "2026-09-29T20:52:11Z",
                        "exitCode" : "FAILED",
                        "exitMessage" : "Execution became stale",
                        "jobExecutionId" : 4127,
                        "kind" : "BATCH_JOB",
                        "name" : "RawZoneReplayJob",
                        "readCount" : 412000,
                        "request" : {
                          "id" : "0192f5a0-1b2c-7d3e-8f4a-5b6c7d8e9f0a",
                          "requestedBy" : "user:operator",
                          "type" : "replay"
                        },
                        "restartable" : true,
                        "runId" : "job:4127",
                        "skipCount" : 950,
                        "startedAt" : "2026-09-29T20:40:02Z",
                        "status" : "FAILED",
                        "writeCount" : 411050
                      }, {
                        "batchCount" : 60,
                        "duplicateCount" : 0,
                        "durationMs" : 59870,
                        "endedAt" : "2026-09-29T21:18:59.870Z",
                        "kind" : "STREAM",
                        "name" : "gtfs-rt-vehicle-position",
                        "readCount" : 7272,
                        "runId" : "stream:gtfs-rt-vehicle-position:2026-09-29T21:18Z",
                        "skipCount" : 4,
                        "startedAt" : "2026-09-29T21:18:00Z",
                        "status" : "COMPLETED_WITH_SKIPS",
                        "writeCount" : 7268
                      } ],
                      "nextCursor" : "eyJ2IjoxLCJrIjpbXX0"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/PagedResponseJobRunResponse"
                }
              }
            },
            "description" : "A page of runs started in the window (default 1 hour, at most 24 hours)"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Invalid request"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authentication required"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Too many requests"
          },
          "503" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Service temporarily unavailable"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Runs of the ETL, batch jobs and minutes of micro-batches, newest first (viewer)",
        "tags" : [ "job-run-controller" ]
      },
      "post" : {
        "operationId" : "requestJobRun",
        "parameters" : [ {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RunJobBody"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "request" : {
                    "description" : "request",
                    "value" : {
                      "id" : "0192f6c3-4d5e-7f6a-8b9c-0d1e2f3a4b5c",
                      "jobName" : "OtpScorecardJob",
                      "kind" : "RUN",
                      "parameters" : {
                        "serviceDates" : "2026-09-27+2026-09-28"
                      },
                      "requestedAt" : "2026-09-29T21:20:03Z",
                      "requestedBy" : "user:operator",
                      "status" : "PENDING"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/JobRequestResponse"
                }
              }
            },
            "description" : "The request is stored; etl-batch starts the job within 5 seconds. Location is the request"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Invalid request"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authentication required"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Access denied"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "job-not-allowed, or idempotency-key-reused"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Too many requests"
          },
          "503" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Service temporarily unavailable"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Run a job now (operator, Idempotency-Key)",
        "tags" : [ "job-request-controller" ]
      }
    },
    "/api/v1/etl/jobs/summary" : {
      "get" : {
        "operationId" : "getJobSummary",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "bucket",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "summary" : {
                    "description" : "summary",
                    "value" : {
                      "batchJobs" : {
                        "completed" : 14,
                        "failed" : 1,
                        "running" : 1,
                        "stopped" : 0
                      },
                      "bucket" : "1m",
                      "from" : "2026-09-29T20:20:00Z",
                      "stream" : [ {
                        "points" : [ {
                          "batches" : 60,
                          "bucketStart" : "2026-09-29T21:18:00Z",
                          "duplicate" : 0,
                          "failedBatches" : 0,
                          "p95BatchMs" : 212,
                          "read" : 7272,
                          "skipped" : 4,
                          "written" : 7268
                        } ],
                        "source" : "GTFS_RT_VEHICLE_POSITION"
                      } ],
                      "to" : "2026-09-29T21:20:00Z"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/JobSummaryResponse"
                }
              }
            },
            "description" : "A series per stream source with every bucket of the window, zeros included"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Invalid request"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authentication required"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Too many requests"
          },
          "503" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Service temporarily unavailable"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Read, written and skipped records per source over time, and batch jobs by outcome (viewer)",
        "tags" : [ "job-run-controller" ]
      }
    },
    "/api/v1/etl/jobs/{runId}" : {
      "get" : {
        "operationId" : "getJobRun",
        "parameters" : [ {
          "in" : "path",
          "name" : "runId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "batchJob" : {
                    "description" : "batchJob",
                    "value" : {
                      "endedAt" : "2026-09-29T20:52:11Z",
                      "exitCode" : "FAILED",
                      "exitMessage" : "Execution became stale",
                      "jobExecutionId" : 4127,
                      "kind" : "BATCH_JOB",
                      "links" : {
                        "logs" : "http://localhost:3000/explore?left=%7B%7D",
                        "trace" : "http://localhost:3000/explore?left=%7B%7D"
                      },
                      "name" : "RawZoneReplayJob",
                      "parameters" : [ {
                        "identifying" : true,
                        "name" : "replayRequestId",
                        "type" : "java.lang.String",
                        "value" : "0192f5a0-1b2c-7d3e-8f4a-5b6c7d8e9f0a"
                      } ],
                      "request" : {
                        "id" : "0192f5a0-1b2c-7d3e-8f4a-5b6c7d8e9f0a",
                        "requestedBy" : "user:operator",
                        "type" : "replay"
                      },
                      "restartable" : true,
                      "runId" : "job:4127",
                      "startedAt" : "2026-09-29T20:40:02Z",
                      "status" : "FAILED",
                      "steps" : [ {
                        "batchId" : "0192f5a1-7c1e-7d3a-9b2c-4e5f6a7b8c9d",
                        "commitCount" : 823,
                        "filterCount" : 0,
                        "processSkipCount" : 950,
                        "readCount" : 412000,
                        "readSkipCount" : 0,
                        "rollbackCount" : 1,
                        "status" : "FAILED",
                        "stepExecutionId" : 99121,
                        "stepName" : "replayRecords",
                        "writeCount" : 411050,
                        "writeSkipCount" : 0
                      } ]
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/JobRunDetailResponse"
                }
              }
            },
            "description" : "The run"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Invalid request"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authentication required"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "No such run, or an id that is not of the form job:<n> or stream:<listener>:<minute>"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Too many requests"
          },
          "503" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Service temporarily unavailable"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "One run in full: parameters, steps and request of a batch job, micro-batches of a stream run",
        "tags" : [ "job-run-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}